Remove webpack plugin. Rename components. Add else-if for sprite sheet selection
