+++ package.json
@@ -47,7 +47,6 @@
     "babel-jest": "^26.3.0",
     "babel-loader": "^8.1.0",
     "babel-preset-vue": "^2.0.2",
-    "clean-webpack-plugin": "^3.0.0",
     "eslint": "^7.7.0",
     "eslint-plugin-vue": "^6.2.2",
     "eslint-teamcity": "^3.0.1",
