+++ src/components/slds-svg/action-sprite.vue
