+++ src/components/slds-svg/custom-sprite.vue
