+++ src/components/slds-svg/doctype-sprite.vue
