+++ src/components/slds-svg/standard-sprite.vue
