+++ src/components/slds-svg/utility-sprite.vue
