+++ vue.config.js
@@ -1,12 +1,5 @@
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
-
 module.exports = {
   css: {
     extract: false
-  },
-  configureWebpack: {
-    plugins: [
-      new CleanWebpackPlugin()
-    ]
   }
 }
